extract crate-related validation into sub-crate
