+++ crates/crates_io_database/Cargo.toml
@@ -12,6 +12,7 @@ bon = "=3.8.1"
 chrono = { version = "=0.4.42", default-features = false, features = ["serde"] }
 crates_io_diesel_helpers = { path = "../crates_io_diesel_helpers" }
 crates_io_index = { path = "../crates_io_index" }
+crates_io_validation = { path = "../crates_io_validation" }
 diesel = { version = "=2.3.4", features = ["serde_json", "chrono", "numeric"] }
 diesel-async = { version = "=0.7.4", features = ["postgres"] }
 diesel_full_text_search = "=2.3.0"
@@ -24,7 +25,6 @@ serde_json = "=1.0.145"
 sha2 = "=0.10.9"
 thiserror = "=2.0.17"
 tracing = "=0.1.43"
-unicode-xid = "=0.2.6"
 utoipa = { version = "=5.4.0", features = ["chrono"] }
 
 [dev-dependencies]
