+++ crates/crates_io_validation/README.md
@@ -1,0 +1,6 @@
+# crates_io_validation
+
+This package contains some low-level validation functions that are used across
+the code base.
+
+These are mostly in their own crate to be shared with docs.rs.
